# apps/backend/docs/swagger/swagger.yaml
basePath: /
definitions:
...
  http.InvitationRequest:
    properties:
      email:
        type: string
      role:
        enum:
        - admin
        - member
        - viewer
        type: string
    required:
    - email
    - role
    type: object
...
  http.RoleRequest:
    properties:
      role:
        enum:
        - admin
        - member
        - viewer
        type: string
    required:
    - role
    type: object
...
  http.TokenRequest:
    properties:
      token:
        maxLength: 128
        type: string
    required:
    - token
    type: object
...
  response.APIError:
    properties:
      code:
        type: string
      details:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
  response.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.APIError'
      message:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
  response.Meta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is the Logify backend API server.
  title: Logify API
  version: "1.0"
paths:
...
  /v1/invitations/accept:
    post:
      consumes:
      - application/json
      parameters:
      - description: Invitation token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.TokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Accept a team invitation
      tags:
      - teams
  /v1/invitations/preview:
    post:
      consumes:
      - application/json
      parameters:
      - description: Invitation token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.TokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Preview a team invitation
      tags:
      - teams
...
  /v1/projects/{id}/invitations:
    post:
      consumes:
      - application/json
      parameters:
      - description: Project UUID
        in: path
        name: id
        required: true
        type: string
      - description: Email and project role
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.InvitationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Invite a project team member
      tags:
      - teams
  /v1/projects/{id}/invitations/{invitationId}:
    delete:
      parameters:
      - description: Project UUID
        in: path
        name: id
        required: true
        type: string
      - description: Invitation UUID
        in: path
        name: invitationId
        required: true
        type: string
      responses:
        "204":
          description: No content
      security:
      - BearerAuth: []
      summary: Cancel a project invitation
      tags:
      - teams
  /v1/projects/{id}/members/{userId}:
    delete:
      parameters:
      - description: Project UUID
        in: path
        name: id
        required: true
        type: string
      - description: User UUID
        in: path
        name: userId
        required: true
        type: string
      responses:
        "204":
          description: No content
      security:
      - BearerAuth: []
      summary: Remove or leave a project team
      tags:
      - teams
    patch:
      consumes:
      - application/json
      parameters:
      - description: Project UUID
        in: path
        name: id
        required: true
        type: string
      - description: User UUID
        in: path
        name: userId
        required: true
        type: string
      - description: Admin, Member, or Viewer
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.RoleRequest'
      responses:
        "204":
          description: No content
      security:
      - BearerAuth: []
      summary: Change project member role
      tags:
      - teams
  /v1/projects/{id}/team:
    get:
      parameters:
      - description: Project UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get project team
      tags:
      - teams
...
securityDefinitions:
  APIKeyAuth:
    description: Project API key, restricted to sending logs.
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: Enter Bearer followed by your access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"